+++ app/api/paypal/cancel-subscription/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 import { cancelSubscription } from "@/lib/paypal";
-import supabaseClient from "@/database/supabase/supabase";
+import supabaseClient from "@/database/supabase/client";
 
 export async function POST(req) {
   try {
