+++ app/api/paypal/create-subscription/route.ts
@@ -1,7 +1,7 @@
 // app/api/paypal/create-subscription/route.js
 import { NextResponse } from "next/server";
 import { createSubscription } from "@/lib/paypal";
-import supabaseClient from "@/database/supabase/supabase";
+import supabaseClient from "@/database/supabase/client";
 
 export async function POST(req) {
   try {
