+++ app/api/stripe/webhook/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import Stripe from "stripe";
 import { headers } from "next/headers";
-import supabaseClient from "@/database/supabase/supabase";
+import supabaseClient from "@/database/supabase/client";
 import { publishToQueue } from "@/lib/upstash";
 
 // Initialize Stripe with your secret key
