+++ components/admin/AlertHoursManager.tsx
@@ -6,7 +6,7 @@ import { Dialog, DialogContent } from "@/components/ui/dialog";
 import { Button } from "@/components/ui/button";
 import { PlusCircle, Loader2, Clock } from "lucide-react";
 import { AlertHours } from "@/hooks/useAlertHours";
-import supabaseClient from "@/database/supabase/supabase";
+import supabaseClient from "@/database/supabase/client";
 import { useToast } from "@/hooks/use-toast";
 import { cn } from "@/lib/utils";
 import { useTheme } from "@/context/theme-context";
