+++ components/admin/alert-hours/AlertHoursList.tsx
@@ -13,7 +13,7 @@ import { Input } from "@/components/ui/input";
 import { Trash2, Clock, Search, ChevronLeft, ChevronRight } from "lucide-react";
 import { cn } from "@/lib/utils";
 import { useTheme } from "@/context/theme-context";
-import supabaseClient from "@/database/supabase/supabase";
+import supabaseClient from "@/database/supabase/client";
 import { useState, useEffect } from "react";
 import {
   Select,
