+++ components/admin/alert-hours/InstrumentSearch.tsx
@@ -3,7 +3,7 @@ import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 import { useTheme } from "@/context/theme-context";
-import supabaseClient from "@/database/supabase/supabase";
+import supabaseClient from "@/database/supabase/client";
 import { Shuffle } from "lucide-react";
 
 interface InstrumentSearchProps {
