--- database/supabase/supabase.js
@@ -1,13 +1,0 @@
-// filepath: /path/to/supabaseClient.ts
-import { createClient } from "@supabase/supabase-js";
-
-const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
-const supabaseKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
-
-if (!supabaseUrl || !supabaseKey) {
-  throw new Error("Missing Supabase environment variables");
-}
-
-const supabaseClient = createClient(supabaseUrl, supabaseKey);
-
-export default supabaseClient;
